+++ client/src/main/java/client/scenes/TaskDetailsCtrl.java
@@ -5,7 +5,13 @@ import client.utils.ServerUtils;
 import com.google.inject.Inject;
 import commons.Task;
 import javafx.fxml.FXML;
+import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.Label;
+import javafx.scene.image.Image;
+import javafx.scene.image.ImageView;
+
+import java.util.Optional;
 
 public class TaskDetailsCtrl {
     private final MainCtrlTalio mainCtrl;
@@ -17,6 +23,10 @@ public class TaskDetailsCtrl {
     private Label title;
     @FXML
     private Label description;
+    @FXML
+    private ImageView editIcon;
+    @FXML
+    private ImageView deleteIcon;
 
     /**
      * Constructor for the task details
@@ -27,6 +37,20 @@ public class TaskDetailsCtrl {
     public TaskDetailsCtrl(ServerUtils server, MainCtrlTalio mainCtrl) {
         this.server = server;
         this.mainCtrl = mainCtrl;
+    }
+
+    /**
+     * This is called only once by the FXML builder,
+     * after FXML components are initialized.
+     */
+    public void initialize() {
+        Image editIcon = new Image(getClass()
+                .getResourceAsStream("/client/images/editicon.png"));
+        this.editIcon.setImage(editIcon);
+
+        Image deleteIcon = new Image(getClass()
+                .getResourceAsStream("/client/images/deleteicon.png"));
+        this.deleteIcon.setImage(deleteIcon);
     }
 
     /**
@@ -68,8 +92,25 @@ public class TaskDetailsCtrl {
             ErrorUtils.alertError("No task to delete!");
             exit();
         }
-        server.deleteTask(task);
-        exit();
+
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.setTitle("Delete Task?");
+        alert.setHeaderText("Delete Task");
+        alert.setContentText("Are you sure you want to delete this task?");
+
+        ButtonType confirmButton = new ButtonType("Confirm");
+        ButtonType cancelButton = new ButtonType("Cancel");
+
+        // Remove the default buttons and add Confirm and Cancel buttons
+        alert.getButtonTypes().setAll(confirmButton, cancelButton);
+
+        Optional<ButtonType> result = alert.showAndWait();
+
+        // Check the user's response and perform the desired action
+        if (result.isPresent() && result.get() == confirmButton) {
+            server.deleteTask(task);
+            exit();
+        }
     }
 
 }
